v2: move serialized resource from function to host

+++ fx-types/abi_function_resources.capnp
@@ -1,5 +1,5 @@
 @0xe3f3e4f9813db34e;
 
 struct FunctionResponse {
-    status @0 :UInt8;
+    status @0 :UInt16;
 }
